requirements.txt
Django==1.6
Django-tinymce-filebrowser==0.2.1
MySQL-python
Pillow==2.2.2
South==0.8.4
django-appconf==0.6
django-compressor==1.3
django-registration==1.0
django-taggit==0.11.2
django-tinymce==1.5.2
gunicorn==18.0
requests==2.1.0
six==1.5.2
sorl-thumbnail==11.12
wsgiref==0.1.2
universal-analytics-python==0.2.4